requirements.txt
fonttools
requests
Pillow
